backend/requirements.txt
fastapi 
uvicorn 
oracledb 
pandas 
numpy 
requests 
python-dotenv